Rename folder SEGD -> segdio
