+++ segdio/__init__.py
